+++ packages/yoteams-build-core/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [*Unreleased*]- <*2021-09-13*>
+## [*1.2.1*]- <*2021-09-14*>
 
 ### Added